pyproject.toml
[tool.poetry]
name = "email-vault-backend"
version = "0.1.0"
description = ""
authors = ["araheeldev <araheeldev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
flask-jwt-extended = "^4.6.0"
werkzeug = "^3.0.6"
python-dotenv = "^1.0.0"
pyotp = "^2.9.0"
qrcode = "^8.0"
pymysql = "1.1.0"      
sqlalchemy = "^2.0.36"
rococo = "^1.0.23"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
